common: malloc: cheri: Change pointer casting to uintptr_t
Change pointer casting from ulong to uintptr_t to match the
pointer type and cast the pointer to ulong for extracting the address
for the architectures the address at the lower part of the pointer type.
(e.g. CHERI).

+++ common/dlmalloc.c
@@ -571,17 +571,17 @@ static mbinptr av_[NAV * 2 + 2] = {
 static void malloc_init(void);
 #endif
 
-ulong mem_malloc_start = 0;
-ulong mem_malloc_end = 0;
-ulong mem_malloc_brk = 0;
+uintptr_t mem_malloc_start = 0;
+uintptr_t mem_malloc_end = 0;
+uintptr_t mem_malloc_brk = 0;
 
 static bool malloc_testing;	/* enable test mode */
 static int malloc_max_allocs;	/* return NULL after this many calls to malloc() */
 
 void *sbrk(ptrdiff_t increment)
 {
-	ulong old = mem_malloc_brk;
-	ulong new = old + increment;
+	uintptr_t old = mem_malloc_brk;
+	uintptr_t new = old + increment;
 
 	if ((new < mem_malloc_start) || (new > mem_malloc_end))
 		return (void *)MORECORE_FAILURE;
@@ -600,7 +600,7 @@ void *sbrk(ptrdiff_t increment)
 
 void mem_malloc_init(ulong start, ulong size)
 {
-	mem_malloc_start = (ulong)map_sysmem(start, size);
+	mem_malloc_start = (uintptr_t)map_sysmem(start, size);
 	mem_malloc_end = mem_malloc_start + size;
 	mem_malloc_brk = mem_malloc_start;
 
@@ -608,8 +608,8 @@ void mem_malloc_init(ulong start, ulong size)
 	malloc_init();
 #endif
 
-	debug("using memory %#lx-%#lx for malloc()\n", mem_malloc_start,
-	      mem_malloc_end);
+	debug("using memory %#lx-%#lx for malloc()\n", (ulong)mem_malloc_start,
+	      (ulong)mem_malloc_end);
 #if CONFIG_IS_ENABLED(SYS_MALLOC_CLEAR_ON_INIT)
 	memset((void *)mem_malloc_start, 0x0, size);
 #endif
@@ -1966,7 +1966,7 @@ Void_t* mEMALIGn_impl(alignment, bytes) size_t alignment; size_t bytes;
     /*
      * m might not be the same as before. Validate that the previous value of
      * extra still works for the current value of m.
-     * If (!m), extra2=alignment so 
+     * If (!m), extra2=alignment so
      */
     if (m) {
       extra2 = alignment - (((unsigned long)(m)) % alignment);
